refactor(scan): move plustek-specific modules & tests
